// src/app/claim-processing/manual-claim-processing/bill-line-messages/model/blmessages.model.ts
import {BillLineMessages} from './bill-line-messages.model';

export class Blmessages {
  aa3cmnd = '';
  aa3mem1 = '';
  aa3mem2 = '';
  a02mem3 = '';
  aa3cno1 = '';
  aa3cno2 = '';
  aa3cno3 = '';
  aa3cno4 = '';
  aa3cno5 = '';
  aa3cno6 = '';
  aa3com1 = '';
  aa3com2 = '';
  aa3hdr = '';
  aa3name = '';
  aa3pln1 = '';
  aa3eff1 = '';
  aa3trm1 = '';
  aa3memn = '';
  aa3assc = '';
  aa3st = '';
  aa3pln2 = '';
  aa3eff2 = '';
  aa3trm2 = '';
  aa3dob = '';
  aa3pln3 = '';
  aa3eff3 = '';
  aa3trm3 = '';
  aa3pln4 = '';
  aa3eff4 = '';
  aa3trm4 = '';
  aa3err = '';
  aa3claim = '';
  aa3qrnbr = '';
  aa3insr = '';
  billlinemessagetab: BillLineMessages[] = [];
}